requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.15
django-allauth==64.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozendict==2.4.4
html5lib==1.1
idna==3.7
inflection==0.5.1
lxml==5.3.0
multitasking==0.0.11
numpy==2.0.1
packaging==24.1
pandas==2.2.2
peewee==3.17.6
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
webencodings==0.5.1
yfinance==0.2.41
